requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
boto3==1.12.47
botocore==1.15.47
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-dropbox-storage==0.4.1
django-filter==2.2.0
django-grappelli==2.14.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-storages==1.9.1
django-taggit==1.2.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-writable-nested==0.6.0
dropbox==10.1.0
gunicorn==20.0.4
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
